x_days_unattached_disks_aws/requirements.txt
boto3
python-dotenv
requests
datetime
awscli

